Fixed conflicts with RStudio 1.3
I don't know if the changes to the CSS code make this conflict with older versions of RStudio so I'm uploading as a separate file. I believe all issues from RStudio 1.3 have been fixed with the exception of some buttons in the menu having a white space around them. I'm working on this fix.